pyk/src/tests/unit: rename mock file
